Async events to event loop thread (#45)
Async provides a mechanism to marshal an event from a thread to the main thread. This is exists primarily to support calling event loop functions from a thread, though there are other use cases.

Event loop functions must be called from the same thread the event loop is running on, by default the main thread. Failure to so will result in bad undefined things eventually happening.